feat: restructure Whisper service; modularize code into routers, implement model loading and transcription logic, and enhance healthcheck functionality
